refactor: ステータス更新機能をManageThreadDropDownに移動
- ThreadStatusBadgeを表示専用コンポーネントに変更
- ManageThreadDropDownにステータス変更メニューを追加
- WIP/CLOSEDの変更をドロップダウンから実行可能
