+++ src/features/threadDetail/ThreadInformation/ManageThreadDropDown/index.tsx
@@ -10,16 +10,20 @@ import {
 } from "@/shared/ui/dropdown-menu";
 import { DropdownMenuItemWithIcon } from "@/shared/ui/dropdown-menu-item-with-icon";
 import { Tooltip } from "@/shared/ui/Tooltip";
-import { Archive, MoreHorizontal, PlaneTakeoff, Trash2 } from "lucide-react";
+import { trpc } from "@/trpc/client";
+import { ThreadStatus } from "@prisma/client";
+import { Archive, LockKeyhole, MoreHorizontal, Pencil, PlaneTakeoff, Trash2 } from "lucide-react";
 import Link from "next/link";
 import { useRouter } from "next/navigation";
 import { useState } from "react";
+import { toast } from "sonner";
 
 type Props = {
   threadId: string;
   threadTitle: string | null;
   includeIsArchived: boolean;
   onClickToggleDisplayArchiveButton: () => void;
+  status: ThreadStatus;
 };
 
 export function ManageThreadDropDown({
@@ -27,12 +31,31 @@ export function ManageThreadDropDown({
   threadTitle,
   includeIsArchived,
   onClickToggleDisplayArchiveButton,
+  status,
 }: Props) {
   const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
   const router = useRouter();
+  const utils = trpc.useUtils();
+
+  const { mutate: updateStatus, isPending } = trpc.thread.updateThreadStatus.useMutation({
+    onSuccess: () => {
+      toast.success("ステータスを更新しました");
+      utils.thread.getThreadInfo.invalidate({ id: threadId });
+      utils.thread.listThreadsByCurrentUser.invalidate();
+    },
+    onError: (error) => {
+      toast.error(`ステータスの更新に失敗しました: ${error.message}`);
+    },
+  });
 
   const handleDeleteSuccess = () => {
     router.push(urls.dashboard);
+  };
+
+  const handleStatusChange = (newStatus: ThreadStatus) => {
+    if (newStatus !== status) {
+      updateStatus({ id: threadId, status: newStatus });
+    }
   };
 
   return (
@@ -58,6 +81,23 @@ export function ManageThreadDropDown({
             text={includeIsArchived ? "アーカイブの非表示" : "アーカイブの表示"}
             onClick={onClickToggleDisplayArchiveButton}
           />
+          <DropdownMenuSeparator />
+          {status !== "WIP" && (
+            <DropdownMenuItemWithIcon
+              icon={Pencil}
+              text="WIPに変更する"
+              onClick={() => handleStatusChange("WIP")}
+              disabled={isPending}
+            />
+          )}
+          {status !== "CLOSED" && (
+            <DropdownMenuItemWithIcon
+              icon={LockKeyhole}
+              text="Closedに変更する"
+              onClick={() => handleStatusChange("CLOSED")}
+              disabled={isPending}
+            />
+          )}
           <DropdownMenuSeparator />
           <DropdownMenuItemWithIcon
             icon={Trash2}
